pyproject.toml
[project]
name = "api_usuarios"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "LocatelliBryann", email = "bryannlocatelli@gmail.com"},
]
dependencies = ["django>=5.2.1", "djangorestframework>=3.16.0", "mysqlclient>=2.2.7"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
